+++ src/main/java/net/litetex/capes/config/AnimatedTexturesHandling.java
@@ -1,0 +1,8 @@
+package net.litetex.capes.config;
+
+public enum AnimatedTexturesHandling
+{
+	ON,
+	FROZEN,
+	OFF
+}