+++ src/main/java/net/litetex/capes/config/Config.java
@@ -16,6 +16,7 @@ public class Config
 	// NOTE: Default/Minecraft is always active
 	private Set<String> activeProviderIds;
 	private boolean enableElytraTexture;
+	private AnimatedTexturesHandling animatedTexturesHandling = AnimatedTexturesHandling.ON;
 	private List<CustomProviderConfig> customProviders = List.of();
 	
 	// Advanced/Debug options
@@ -27,6 +28,7 @@ public class Config
 		this.setCurrentPreviewProviderId(null);
 		this.setActiveProviderIds(List.of(MinecraftCapesCapeProvider.ID, OptiFineCapeProvider.ID));
 		this.setEnableElytraTexture(true);
+		this.setAnimatedTexturesHandling(AnimatedTexturesHandling.ON);
 		
 		this.setValidateProfile(null);
 		this.setLoadThrottleSuppressSec(null);
@@ -69,6 +71,16 @@ public class Config
 	public void setEnableElytraTexture(final boolean enableElytraTexture)
 	{
 		this.enableElytraTexture = enableElytraTexture;
+	}
+	
+	public AnimatedTexturesHandling getAnimatedTexturesHandling()
+	{
+		return this.animatedTexturesHandling;
+	}
+	
+	public void setAnimatedTexturesHandling(final AnimatedTexturesHandling animatedTexturesHandling)
+	{
+		this.animatedTexturesHandling = animatedTexturesHandling;
 	}
 	
 	public List<CustomProviderConfig> getCustomProviders()